SunoSongUploader/requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
requests
SunoAI
pillow